reworked database model some more based off feedback. started adding some of the logging module. Also started working on an async io task helper but it still needs some work

+++ Cybermancy.Core/Extensions/RewardExtension.cs
@@ -11,16 +11,16 @@ namespace Cybermancy.Core.Extensions
 {
     public static class RewardExtension
     {
-        public static int GetXpNeeded(this Reward reward)
+        public static uint GetXpNeeded(this Reward reward)
         {
-            var level = reward.RewardLevel - 2;
+            var level = (int)reward.RewardLevel - 2;
             return level switch
             {
                 0 => reward.Guild.LevelSettings.Base,
-                < 0 => 0,
-                _ => reward.Guild.LevelSettings.Base + ((int)Math.Round(reward.Guild.LevelSettings.Base *
+                < 0 => (uint)0,
+                _ => reward.Guild.LevelSettings.Base + ((uint)Math.Round(reward.Guild.LevelSettings.Base *
                                                                          (reward.Guild.LevelSettings.Modifier / 100.0) *
-                                                                         level) * level)
+                                                                         (uint)level) * (uint)level)
             };
         }
